git-svn-id: http://dminfo:8686/svn/repository/models/trunk@372 cdc4813b-4270-ec4f-936e-925f93a782c2

+++ dsm2/src/ptm/native/DWR_DMS_PTM_PTMFixedData.c
@@ -425,7 +425,7 @@ JNIEXPORT jfloatArray JNICALL Java_DWR_DMS_PTM_PTMFixedData_getChannelXSectionDi
 JNIEXPORT jfloat JNICALL Java_DWR_DMS_PTM_PTMFixedData_getReservoirArea
 (JNIEnv * env, jobject thisOne, jint i){
   int reservoirNumber = (int)i;
-  jfloat area = (jint) getReservoirArea(&reservoirNumber);
+  jfloat area = getReservoirArea(&reservoirNumber);
   return area;
 }
 
@@ -437,7 +437,7 @@ JNIEXPORT jfloat JNICALL Java_DWR_DMS_PTM_PTMFixedData_getReservoirArea
 JNIEXPORT jfloat JNICALL Java_DWR_DMS_PTM_PTMFixedData_getReservoirBottomElevation
 (JNIEnv * env, jobject thisOne, jint i){
   int reservoirNumber = (int)i;
-  jfloat bottomElevation = (jint) getReservoirArea(&reservoirNumber);
+  jfloat bottomElevation = getReservoirArea(&reservoirNumber);
   return bottomElevation;
 }
